+++ packages/terra-form-select/tests/jest/Menu.test.jsx
@@ -1,14 +1,15 @@
 import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl } from 'terra-enzyme-intl';
 import Menu from '../../src/_Menu';
 import Option from '../../src/_Option';
-import intlContexts from './intl-context-setup';
 
 describe('Menu', () => {
   it('should render a default Menu', () => {
     const liveRegion = { current: document.createElement('div') };
 
     const menu = <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} variant="default" value="value" />;
-    const wrapper = shallow(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -25,7 +26,7 @@ describe('Menu', () => {
 
     jest.useFakeTimers();
 
-    const wrapper = shallow(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
 
     wrapper.setState({ searchValue: 'value' });
 
@@ -45,7 +46,7 @@ describe('Menu', () => {
       </Menu>
     );
 
-    const wrapper = shallow(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
 
     jest.useFakeTimers();
 
@@ -54,6 +55,64 @@ describe('Menu', () => {
     jest.advanceTimersByTime(500);
 
     expect(wrapper).toMatchSnapshot();
-    expect(liveRegion.current.innerText).toEqual('No matching results for "asdf"');
+  });
+
+  it('should not error when visuallyHiddenComponent is not provided', () => {
+    const mockIntl = { formatMessage: id => (`No Results for ${id.id}`) };
+
+    const menu = (
+      <Menu onSelect={() => {}} intl={mockIntl} variant="default" value="value" searchValue="asdf">
+        <Option value="value" display="display" />
+      </Menu>
+    );
+
+    const wrapper = shallowWithIntl(menu);
+
+    jest.useFakeTimers();
+
+    wrapper.setState({ hasNoResults: true });
+
+    jest.advanceTimersByTime(500);
+    expect(true).toEqual(true);
+  });
+
+  it('should not error when visuallyHiddenComponent has null for current', () => {
+    const liveRegion = { current: null };
+    const mockIntl = { formatMessage: id => (`No Results for ${id.id}`) };
+
+    const menu = (
+      <Menu onSelect={() => {}} intl={mockIntl} visuallyHiddenComponent={liveRegion} variant="default" value="value" searchValue="asdf">
+        <Option value="value" display="display" />
+      </Menu>
+    );
+
+    const wrapper = shallowWithIntl(menu);
+
+    jest.useFakeTimers();
+
+    wrapper.setState({ hasNoResults: true });
+
+    jest.advanceTimersByTime(500);
+    expect(true).toEqual(true);
+  });
+
+  it('should render a Menu with the clear option', () => {
+    const liveRegion = { current: document.createElement('div') };
+
+    jest.useFakeTimers();
+
+    const menu = (
+      <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} variant="combobox" clearOptionDisplay="-Select-" searchValue="">
+        <Option value="value" display="display" />
+      </Menu>
+    );
+
+    jest.useFakeTimers();
+
+    const wrapper = shallowWithIntl(menu);
+
+    jest.runOnlyPendingTimers();
+
+    expect(wrapper).toMatchSnapshot();
   });
 });