+++ packages/terra-form-select/tests/jest/SelectField.test.jsx
@@ -1,5 +1,8 @@
 import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl } from 'terra-enzyme-intl';
 import SelectField from '../../src/SelectField';
+
 
 it('should render a default SelectField component', () => {
   const select = (
@@ -12,7 +15,7 @@ it('should render a default SelectField component', () => {
     </SelectField>
   );
 
-  expect(shallow(select)).toMatchSnapshot();
+  expect(shallowWithIntl(select)).toMatchSnapshot();
 });
 
 it('should render a default SelectField component with attributes spread onto the select', () => {
@@ -26,5 +29,19 @@ it('should render a default SelectField component with attributes spread onto th
     </SelectField>
   );
 
-  expect(shallow(select)).toMatchSnapshot();
+  expect(shallowWithIntl(select)).toMatchSnapshot();
+});
+
+it('should render a multiple SelectField component that limits the selection count', () => {
+  const select = (
+    <SelectField help="Help Text" label="Label" maxSelectionCount={2} variant="multiple" defaultValue="blue" selectId="select-id" selectAttrs={{ onClick: () => {} }}>
+      <SelectField.Option value="blue" display="Blue" />
+      <SelectField.Option value="green" display="Green" />
+      <SelectField.Option value="purple" display="Purple" />
+      <SelectField.Option value="red" display="Red" />
+      <SelectField.Option value="violet" display="Violet" />
+    </SelectField>
+  );
+
+  expect(shallowWithIntl(select)).toMatchSnapshot();
 });