+++ packages/terra-progress-bar/src/terra-dev-site/doc/example/ProgressBarSize.jsx
@@ -4,19 +4,19 @@ import ProgressBar from 'terra-progress-bar/lib/ProgressBar';
 
 const ProgressBarSize = () => (
   <div>
-    <ProgressBar heightSize="tiny" valueText="15%" value={15} color="#8ccc62" />
+    <ProgressBar heightSize="tiny" valueText="15%" value={15} />
     <br />
     <br />
-    <ProgressBar heightSize="small" valueText="30%" value={30} color="#8ccc62" />
+    <ProgressBar heightSize="small" valueText="30%" value={30} />
     <br />
     <br />
-    <ProgressBar heightSize="medium" valueText="45%" value={45} color="#8ccc62" />
+    <ProgressBar heightSize="medium" valueText="45%" value={45} />
     <br />
     <br />
-    <ProgressBar heightSize="large" valueText="60%" value={60} color="#8ccc62" />
+    <ProgressBar heightSize="large" valueText="60%" value={60} />
     <br />
     <br />
-    <ProgressBar heightSize="huge" valueText="75%" value={75} color="#8ccc62" />
+    <ProgressBar heightSize="huge" valueText="75%" value={75} />
   </div>
 );