+++ packages/terra-status/src/terra-dev-site/test/status/StatusDefault.test.jsx
@@ -1,11 +1,15 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Status from '../../../Status';
+import styles from '../../doc/example/colors.module.scss';
+
+const cx = classNames.bind(styles);
 
 const simpleText = <div style={{ padding: '5px' }}>Sample text</div>;
 
 const StatusDefault = () => (
   <div>
-    <Status id="status" color="#ff0000" visuallyHiddenText="Status High">{simpleText}</Status>
+    <Status id="status" colorClass={cx(['attention'])} visuallyHiddenText="Status Attention">{simpleText}</Status>
   </div>
 );