+++ packages/terra-text/src/terra-dev-site/doc/example/TextDisplays.jsx
@@ -4,8 +4,11 @@ import Arrange from 'terra-arrange';
 import Text from 'terra-text/lib/Text';
 import classNames from 'classnames/bind';
 import styles from '../../../Text.module.scss';
+import styles2 from './colors.module.scss';
 
 const cx = classNames.bind(styles);
+
+const cx2 = classNames.bind(styles2);
 
 
 const TextDisplays = () => (
@@ -22,7 +25,7 @@ const TextDisplays = () => (
     <br />
     <h3>Font Weight: 300</h3>
     <Arrange
-      fitStart={<Text weight={300} color="#333" fontSize={100}>Aa</Text>}
+      fitStart={<Text weight={300} className={cx2(['default'])} fontSize={100}>Aa</Text>}
       fill={(
         <div style={{ marginLeft: '10px', marginRight: '10px' }}>
           <Text weight={300} fontSize={32}>a b c d e f g h i j k l m n o p q r s t u v w x y z</Text>
@@ -37,7 +40,7 @@ const TextDisplays = () => (
     <br />
     <h3>Font Style: Italic Font Weight: 300</h3>
     <Arrange
-      fitStart={<Text isItalic weight={300} color="#333" fontSize={100}>Aa</Text>}
+      fitStart={<Text isItalic weight={300} className={cx2(['default'])} fontSize={100}>Aa</Text>}
       fill={(
         <div style={{ marginLeft: '10px', marginRight: '10px' }}>
           <Text isItalic weight={300} fontSize={32}>a b c d e f g h i j k l m n o p q r s t u v w x y z</Text>
@@ -52,7 +55,7 @@ const TextDisplays = () => (
     <br />
     <h3>Font Weight: 400</h3>
     <Arrange
-      fitStart={<Text color="#333" fontSize={100}>Aa</Text>}
+      fitStart={<Text className={cx2(['default'])} fontSize={100}>Aa</Text>}
       fill={(
         <div style={{ marginLeft: '10px', marginRight: '10px' }}>
           <Text fontSize={32}>a b c d e f g h i j k l m n o p q r s t u v w x y z</Text>
@@ -67,7 +70,7 @@ const TextDisplays = () => (
     <br />
     <h3>Font Style: Italic</h3>
     <Arrange
-      fitStart={<Text isItalic color="#333" fontSize={100}>Aa</Text>}
+      fitStart={<Text isItalic className={cx2(['default'])} fontSize={100}>Aa</Text>}
       fill={(
         <div style={{ marginLeft: '10px', marginRight: '10px' }}>
           <Text isItalic fontSize={32}>a b c d e f g h i j k l m n o p q r s t u v w x y z</Text>
@@ -82,7 +85,7 @@ const TextDisplays = () => (
     <br />
     <h3>Font Weight: 700</h3>
     <Arrange
-      fitStart={<Text weight={700} color="#333" fontSize={100}>Aa</Text>}
+      fitStart={<Text weight={700} className={cx2(['default'])} fontSize={100}>Aa</Text>}
       fill={(
         <div style={{ marginLeft: '10px', marginRight: '10px' }}>
           <Text weight={700} fontSize={32}>a b c d e f g h i j k l m n o p q r s t u v w x y z</Text>
@@ -97,7 +100,7 @@ const TextDisplays = () => (
     <br />
     <h3>Font Style: Italic Font Weight: 700</h3>
     <Arrange
-      fitStart={<Text isItalic weight={700} color="#333" fontSize={100}>Aa</Text>}
+      fitStart={<Text isItalic weight={700} className={cx2(['default'])} fontSize={100}>Aa</Text>}
       fill={(
         <div style={{ marginLeft: '10px', marginRight: '10px' }}>
           <Text isItalic weight={700} fontSize={32}>a b c d e f g h i j k l m n o p q r s t u v w x y z</Text>