+++ packages/terra-alert/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
+3.8.0 - (March 5, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.7.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 ### Added
 * Added en-AU translations