+++ packages/terra-button/CHANGELOG.md
@@ -3,6 +3,13 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Replaced local keyCode values with keycode-js npm package 
+
+3.6.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.5.0 - (February 12, 2019)
 ------------------