+++ packages/terra-content-container/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+3.3.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+* Removed dependency on terra-heading used in examples
+
 3.2.0 - (February 12, 2019)
 ------------------
 ### Changed