+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -3,7 +3,13 @@ Changelog
 
 Unreleased
 ----------
-### Added
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
+3.5.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 * Added en-AU translations
 
 3.4.0 - (February 12, 2019)