+++ packages/terra-form-field/CHANGELOG.md
@@ -4,6 +4,18 @@ ChangeLog
 Unreleased
 ----------
 ### Added
+* Added logic to allow accessible mapping of error and help text messages when htmlFor prop is provided
+
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
+
+3.4.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
 * Added en-AU translations
 
 3.3.0 - (February 12, 2019)