+++ packages/terra-form-radio/CHANGELOG.md
@@ -3,7 +3,16 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
+3.6.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 ### Added
+* Added stylelint-enable at end of Radio.module.scss
 * Added en-AU translations
 
 3.5.0 - (February 12, 2019)