+++ packages/terra-hyperlink/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Replaced local keyCode values with keycode-js npm package
+
+2.5.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 2.4.0 - (February 12, 2019)
 ------------------