+++ packages/terra-legacy-theme/CHANGELOG.md
@@ -4,6 +4,14 @@ Changelog
 Unreleased
 ----------
 
+2.36.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Added stylelint-enable to LegacyThemeNavigationSideMenu.module.scss
+
 2.35.0 - (February 5, 2019)
 ------------------
 ### Changed