+++ packages/terra-paginator/CHANGELOG.md
@@ -3,6 +3,13 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Replaced local keyCode values with keycode-js npm package
+
+2.7.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 2.6.0 - (February 12, 2019)
 ------------------