+++ packages/terra-progress-bar/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+4.0.0 - (February 26, 2019)
+------------------
+### Breaking Changes
+* Added new colorClass prop and examples of how to use it.
+* Removed color and hasGradient props
+
 3.4.0 - (February 12, 2019)
 ------------------
 ### Changed