+++ packages/terra-scroll/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.4.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.2.0 - (February 12, 2019)
 ------------------
 ### Changed