+++ packages/terra-section-header/CHANGELOG.md
@@ -3,6 +3,14 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Replaced local keyCode values with keycode-js npm package
+
+2.6.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 ### Fixed
 * Correct browser.buttonDown() usage