+++ packages/terra-show-hide/CHANGELOG.md
@@ -3,6 +3,20 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Replaced local keyCode values with keycode-js npm package
+* Update Jest tests to use terra-enzyme-intl
+
+2.5.0 - (March 5, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.4.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 ### Added
 * Added en-AU translations