+++ packages/terra-signature/CHANGELOG.md
@@ -3,6 +3,24 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Removed terra-form-select dependency
+
+2.7.0 - (March 6, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.6.0 - (March 5, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.5.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+* Removed stylelint-disable in Signature.module.scss
 
 2.4.0 - (February 12, 2019)
 ------------------