+++ packages/terra-status-view/CHANGELOG.md
@@ -3,15 +3,29 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
+### Fixed
+* Remove --terra-status-view-message-text-align theme variable to fix alignment within Edge browser.
+
+3.7.0 - (March 5, 2019)
+------------------
+### Fixed
+* Update SVG style to `display: block` to fix SVG size and stop text from shifting downards.
+
+3.6.0 - (February 26, 2019)
+------------------
 ### Added
 * Added en-AU translations
 
 ### Changed
+* Minor dependency version bump
 * Update Thin Constraint test screenshot to be limited to the test size 75px to remove wdio screen resizing. Screen resizing to 75px is not possible with firefox driver.
 
 ### Fixed
 * Status View glyph does not always displays when it should
-=======
+
 3.5.0 - (February 12, 2019)
 ------------------
 ### Changed