+++ packages/terra-toggle-section-header/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.4.1 - (March 5, 2019)
+------------------
+### Changed
+* Patch dependency version bump
+
+2.4.0 - (February 26, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.3.0 - (February 12, 2019)
 ------------------
 ### Changed